refactor: split schema module into package
