+++ relativity/schema/_util.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from dataclasses import Field, field
+
+
+def _dict_field() -> Field:
+    return field(default_factory=dict)
+
+
+__all__ = ["_dict_field"]
